Set numpy array types to float32 and int32 where appropriate

+++ utils/data_splitting.py
@@ -6,7 +6,7 @@ def write_numpy_array_to_file(array, file_path):
 
 
 def create_numpy_array_from_generator(generator, overestimated_shape):
-    array = np.zeros(shape=overestimated_shape)
+    array = np.zeros(shape=overestimated_shape, dtype=np.int32)
     for i, x in enumerate(generator()):
         array[i, :] = x
     return array[:i + 1, :]
